requirements.txt
uvicorn==0.17.6
elasticsearch==8.1.2
aioredis==2.0.1
aiohttp==3.8.1
fastapi[all]==0.75.2
pydantic==1.9.0
aenum==3.1.11
icecream==2.1.2
fastapi-utils== 0.2.1
pytest-asyncio==0.12.0